+++ backend/src/Logitar.Portal.Domain/Templates/UniqueKey.cs
@@ -1,20 +1,19 @@
 ﻿using FluentValidation;
-using Logitar.Portal.Domain.Templates.Validators;
+using Logitar.Identity.Domain.Shared;
 
 namespace Logitar.Portal.Domain.Templates;
 
-public record UniqueKey
+public record UniqueKey // ISSUE #528: replace with Identifier
 {
   public string Value { get; }
 
-  public UniqueKey(string value, string? propertyName = null)
+  public UniqueKey(string value)
   {
     Value = value.Trim();
-    new UniqueKeyValidator(propertyName).ValidateAndThrow(Value);
+    new IdentifierValidator().ValidateAndThrow(Value);
   }
 
-  public static UniqueKey? TryCreate(string? value, string? propertyName = null)
-  {
-    return string.IsNullOrWhiteSpace(value) ? null : new(value, propertyName);
-  }
+  public static UniqueKey? TryCreate(string? value) => string.IsNullOrWhiteSpace(value) ? null : new(value);
+
+  public override string ToString() => Value;
 }