+++ backend/src/Logitar.Portal.Domain/Validators/SlugValidator.cs
@@ -1,14 +1,19 @@
 ﻿using FluentValidation;
-using Logitar.Identity.Domain;
+using FluentValidation.Validators;
 
 namespace Logitar.Portal.Domain.Validators;
 
-internal class SlugValidator : AbstractValidator<string>
+internal class SlugValidator<T> : IPropertyValidator<T, string>
 {
-  public SlugValidator(string? propertyName = null)
+  public string Name { get; } = "SlugValidator";
+
+  public string GetDefaultMessageTemplate(string errorCode)
   {
-    RuleFor(x => x).Must(BeAValidSlug).WithPropertyName(propertyName);
+    return "'{PropertyName}' must be composed of non-empty alphanumeric words separated by hyphens (-).";
   }
 
-  private static bool BeAValidSlug(string slug) => slug.Split('-').All(word => !string.IsNullOrEmpty(word) && word.All(char.IsLetterOrDigit));
+  public bool IsValid(ValidationContext<T> context, string value)
+  {
+    return value.Split('-').All(word => !string.IsNullOrEmpty(word) && word.All(char.IsLetterOrDigit));
+  }
 }